+++ Assets/Scenes/GeneScene.unity
@@ -378,8 +378,16 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 44fa230acc382473caf690be8dd974e2, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  GeneratingText: {fileID: 2087279643}
-  GeneratePanel: {fileID: 1686387551}
+  generatingText: {fileID: 0}
+  originalPlayerInfoPrefab: {fileID: 0}
+  playerInfoContent: {fileID: 0}
+  generatePanel: {fileID: 0}
+  readyPanel: {fileID: 0}
+  geneInputPanel: {fileID: 0}
+  StartButton: {fileID: 0}
+  playerPrefab: {fileID: 0}
+  allPlayerInfo: []
+  playersOrigin: {fileID: 0}
 --- !u!1 &79188256
 GameObject:
   m_ObjectHideFlags: 0
@@ -691,7 +699,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   jyanru: {fileID: 758018778}
-  soloStartButton: {fileID: 1036123132}
 --- !u!114 &244720515
 MonoBehaviour:
   m_ObjectHideFlags: 0
